Extract shared viz backend and harden viz integration
