+++ package.json
@@ -3,9 +3,9 @@
   "private": true,
   "license": "MIT",
   "scripts": {
-    "clean:all": "rm -rf node_modules tooling/satsuma-cli/node_modules tooling/satsuma-core/node_modules tooling/satsuma-viz-model/node_modules tooling/tree-sitter-satsuma/node_modules tooling/satsuma-viz/node_modules tooling/satsuma-lsp/node_modules tooling/vscode-satsuma/node_modules tooling/satsuma-core/dist tooling/satsuma-viz-model/dist tooling/satsuma-cli/dist tooling/satsuma-cli/src/generated tooling/satsuma-viz/dist tooling/satsuma-lsp/dist tooling/vscode-satsuma/dist tooling/tree-sitter-satsuma/build tooling/tree-sitter-satsuma/tree-sitter-satsuma.wasm",
-    "install:all": "npm install && npm --prefix tooling/satsuma-core install && npm --prefix tooling/satsuma-viz-model install && npm --prefix tooling/tree-sitter-satsuma install && cd tooling/tree-sitter-satsuma && tree-sitter build --wasm . -o tree-sitter-satsuma.wasm && cd ../.. && npm --prefix tooling/satsuma-cli install && npm --prefix tooling/satsuma-viz install && npm --prefix tooling/satsuma-viz run build && npm --prefix tooling/satsuma-lsp install && npm --prefix tooling/vscode-satsuma install && npm --prefix tooling/vscode-satsuma run build:server",
-    "ci:all": "npm ci && npm ci --prefix tooling/satsuma-core && npm ci --prefix tooling/satsuma-viz-model && npm ci --prefix tooling/tree-sitter-satsuma && npm ci --ignore-scripts --prefix tooling/satsuma-cli && npm ci --prefix tooling/satsuma-viz && npm ci --prefix tooling/satsuma-lsp && npm ci --prefix tooling/vscode-satsuma",
+    "clean:all": "rm -rf node_modules tooling/satsuma-cli/node_modules tooling/satsuma-core/node_modules tooling/satsuma-viz-model/node_modules tooling/satsuma-viz-backend/node_modules tooling/tree-sitter-satsuma/node_modules tooling/satsuma-viz/node_modules tooling/satsuma-lsp/node_modules tooling/vscode-satsuma/node_modules tooling/satsuma-core/dist tooling/satsuma-viz-model/dist tooling/satsuma-viz-backend/dist tooling/satsuma-cli/dist tooling/satsuma-cli/src/generated tooling/satsuma-viz/dist tooling/satsuma-lsp/dist tooling/vscode-satsuma/dist tooling/tree-sitter-satsuma/build tooling/tree-sitter-satsuma/tree-sitter-satsuma.wasm",
+    "install:all": "npm install && npm --prefix tooling/satsuma-core install && npm --prefix tooling/satsuma-viz-model install && npm --prefix tooling/satsuma-viz-backend install && npm --prefix tooling/tree-sitter-satsuma install && cd tooling/tree-sitter-satsuma && tree-sitter build --wasm . -o tree-sitter-satsuma.wasm && cd ../.. && npm --prefix tooling/satsuma-cli install && npm --prefix tooling/satsuma-viz install && npm --prefix tooling/satsuma-viz run build && npm --prefix tooling/satsuma-lsp install && npm --prefix tooling/vscode-satsuma install && npm --prefix tooling/vscode-satsuma run build:server",
+    "ci:all": "npm ci && npm ci --prefix tooling/satsuma-core && npm ci --prefix tooling/satsuma-viz-model && npm ci --prefix tooling/satsuma-viz-backend && npm ci --prefix tooling/tree-sitter-satsuma && npm ci --ignore-scripts --prefix tooling/satsuma-cli && npm ci --prefix tooling/satsuma-viz && npm ci --prefix tooling/satsuma-lsp && npm ci --prefix tooling/vscode-satsuma",
     "reinstall": "npm run clean:all && npm run install:all",
     "lint": "npm run lint:js && npm run lint:md && npm run lint:yaml && npm run lint:python",
     "lint:js": "eslint .",
