+++ tooling/satsuma-lsp/package.json
@@ -9,14 +9,15 @@
     "satsuma-lsp": "bin/satsuma-lsp.js"
   },
   "scripts": {
-    "prebuild": "npm --prefix ../satsuma-core run build",
+    "prebuild": "npm --prefix ../satsuma-core run build && npm --prefix ../satsuma-viz-backend run build",
     "build": "esbuild src/server.ts --bundle --platform=node --target=node20 --outfile=dist/server.js --sourcemap --banner:js='var __import_meta_url = require(\"url\").pathToFileURL(__filename).href;' --define:import.meta.url=__import_meta_url",
     "watch": "npm run build -- --watch",
     "test": "node --test test/**/*.test.js",
-    "pretest": "npm --prefix ../satsuma-core run build && tsc"
+    "pretest": "npm --prefix ../satsuma-core run build && npm --prefix ../satsuma-viz-backend run build && tsc"
   },
   "dependencies": {
     "@satsuma/core": "file:../satsuma-core",
+    "@satsuma/viz-backend": "file:../satsuma-viz-backend",
     "@satsuma/viz-model": "file:../satsuma-viz-model",
     "vscode-languageserver": "^9.0.1",
     "vscode-languageserver-textdocument": "^1.0.12",
