+++ tooling/satsuma-viz-backend/tsconfig.json
@@ -3,22 +3,18 @@
     "module": "commonjs",
     "moduleResolution": "node",
     "baseUrl": ".",
-    "ignoreDeprecations": "6.0",
     "target": "es2022",
     "types": ["node"],
     "paths": {
-      "web-tree-sitter": ["./src/web-tree-sitter.d.ts"],
       "@satsuma/core": ["./node_modules/@satsuma/core/dist/index.d.ts"],
-      "@satsuma/core/format": ["./node_modules/@satsuma/core/dist/format.d.ts"],
-      "@satsuma/viz-backend": ["./node_modules/@satsuma/viz-backend/dist/index.d.ts"],
-      "@satsuma/viz-backend/workspace-index": ["./node_modules/@satsuma/viz-backend/dist/workspace-index.d.ts"],
-      "@satsuma/viz-backend/viz-model": ["./node_modules/@satsuma/viz-backend/dist/viz-model.d.ts"]
+      "@satsuma/viz-model": ["./node_modules/@satsuma/viz-model/dist/index.d.ts"]
     },
     "strict": true,
     "noUncheckedIndexedAccess": true,
     "outDir": "dist",
     "rootDir": "src",
     "declaration": true,
+    "declarationMap": true,
     "sourceMap": true,
     "esModuleInterop": true,
     "skipLibCheck": true,
