+++ tooling/satsuma-viz/package.json
@@ -10,7 +10,7 @@
     "build": "esbuild src/satsuma-viz.ts --bundle --format=esm --target=es2022 --outfile=dist/satsuma-viz.js --sourcemap --loader:.css=text --minify",
     "watch": "npm run build -- --watch",
     "test": "node --test test/**/*.test.js",
-    "pretest": "tsc --noEmit"
+    "pretest": "tsc --noEmit && npm run build"
   },
   "dependencies": {
     "@satsuma/core": "file:../satsuma-core",
