+++ tooling/vscode-satsuma/package.json
@@ -109,7 +109,8 @@
     "build": "node esbuild.js",
     "watch": "node esbuild.js --watch",
     "build:server": "cd ../satsuma-lsp && npm run build",
-    "test": "npm run test:fixtures && npm run test:golden",
+    "test": "npm run test:unit && npm run test:fixtures && npm run test:golden",
+    "test:unit": "tsc -p src/tsconfig.json && node --test test/**/*.test.js",
     "test:fixtures": "npx vscode-tmgrammar-test -g syntaxes/satsuma.tmLanguage.json 'test/fixtures/*.stm'",
     "test:golden": "npx vscode-tmgrammar-test -g syntaxes/satsuma.tmLanguage.json 'test/golden/*.stm'",
     "test:lsp": "cd ../satsuma-lsp && npm test",
@@ -124,6 +125,7 @@
     "vscode-languageclient": "^9.0.1"
   },
   "devDependencies": {
+    "@types/node": "^25.5.0",
     "@types/vscode": "^1.85.0",
     "esbuild": "^0.27.4",
     "typescript": "^6.0.2",
